requirements.txt
torch==2.1.0
matplotlib==3.8.0
torchvision==0.16.0
Pillow==10.1.0
Flask==3.0.3
